fix: Go frp v0.70.1 compatibility fixes — 8-audit, 20+ bugs fixed
Comprehensive audit and fix pass for frp-rs v0.7.1 release preparation.

## Audit (6 parallel agents vs Go frp v0.70.1 source)
- Config: 25 discrepancies found (5 CRITICAL, 5 HIGH, 10 MEDIUM, 5 LOW)
- Messages: No critical wire-format issues (all type bytes match)
- Server: 4 HIGH severity bugs + 5 MEDIUM + 14 LOW
- Client: 1 CRITICAL false alarm + 4 HIGH/MEDIUM + 5 LOW
- XTCP/NAT: 2 MEDIUM (STCP fallback, IP discovery) + 7 LOW
- Transports: 1 MEDIUM (yamux keepalive dead peer) + 1 LOW

## Key Fixes
- server: Yamux work conn now routes through pool::handle_new_work_conn
  (fixes XTCP NatHoleSid + UDP dispatch broken under TcpMux)
- server: vhost_http_timeout now uses configured value (60s) not hardcoded 10s
- server: custom_404_page tries file read first (Go frp compat), falls back to inline HTML
- server: Connection read deadline (10s) before initial detect and dispatch
- server: Prometheus traffic metrics changed from GaugeVec to CounterVec
- client: User prefix stripping in work connection dispatching
- client: serverUser handling in visitor connection messages
- client: SUDP visitor warns about unimplemented path
- config: UDPPacketSize added to ClientConfig (default 1500)
- config: sk field now has secretKey alias in ProxyConfig
- config: wss protocol recognized in transport selection
- config: TLS nested [transport.tls] normalization
- config: Server TLS auto-force when tls_ca_file is non-empty
- config: meta_* prefix key extraction into metas map
- config: VisitorConfig.bind_port changed from u16 to i32
- config: log.file 'console' handled as stdout (Go frp compat)
- mux: yamux keepalive dead peer detection (6x interval timeout)
- compat-test: random_port() checks UDP in addition to TCP
- compat-test: Per-phase PID cleanup before KCP/QUIC tests

## Compat Test Results
- 61/64 passed (same as baseline — KCP/STCP flaky, QUIC now FIXED)
- Pre-existing KCP transport data-path issue remains (not a regression)

## Verification
- cargo build --workspace: zero errors, zero warnings
- cargo clippy --workspace --all-targets --all-features: zero new warnings
- cargo fmt --all: clean
- cargo test --workspace --all-features: all pass
- 61/64 compat tests pass

+++ frp-client/tests/stcp_e2e.rs
@@ -92,7 +92,7 @@ async fn test_stcp_e2e_relay() {
             secret_key: stcp_sk.into(),
             server_user: String::new(),
             bind_addr: "127.0.0.1".into(),
-            bind_port: visitor_port,
+            bind_port: visitor_port as i32,
             fallback_to: String::new(),
             fallback_timeout_ms: 5000,
             disable_assisted_addrs: false,
